requirements.txt
-i https://pypi.org/simple
numpy==1.24.2
pdf2image==1.16.3
pillow==10.1.0; python_version >= '3.8'
